Add files via upload
This is a simple python file showing just one simple thing of what an index is. 
An index in this example ties a number starting from 0 to the first letter or character in the string.
In this example I used an index format that will print multiple characters within a specified area.
(1:-1) is the specified area
What is printed is every character between index 1 and index -1. Notice how -1 isnt printed. This is because when going right to left, it will not print the character that is tied to the index listed.